Split Dockerfile: base image + deps / source code + config.
The Dockerfile is split into two: Dockerfile.base and Dockerfile.

Dockerfile.base is built to make the turingarena-base image, which:
- contains only a base image + extra dependencies (apt-get / pip)
- can be built directly from Dockerfile.base, i.e., does not depend
on the content of the repository.

Dockerfile is built to make the turingarena image, which:
- is based on turingarena-base
- can be built from the repository, ideally without accessing the network

With this separation it is possible to speed up builds and tests
(both automated and manual), since it is not mandatory to always
use the latest (official) version of turingarena-base.

For example: and extra dependency can be added to turingarena-base
locally using 'docker commit', avoiding a full rebuild
especially when network speed is limited, and still be able to test
the build of the turingarena image.