Add constants to interface.txt
Now you can define constants in `interface.txt`.
To define them you use the following syntax: `const NAME = VALUE`
All constants declarations must be put at the begining of interface.txt,
before any method declaration.
The constants will appear in the generated template, so they can be use
by the person that writes the solution.

For now you can't use constants in interface.txt, but obviously in the
very near future we will support that as well, as also having a method
to get the values of the constants in the evaluator.