Do not initialize a git repository in the current directory
Instead print an error saying that to work on a problem you must
be inside a git repository. This avoids error like launching
turingarena in the wrong directory (~/, for example) and thus
creating unwanted git repositories everywhere.

If the user create a new problem with the command 'turingarena new'
a git repository in the current directory is initialized by default
anyway.